Move pageRenderer to server/render/
